// src/api/generated/model/mgmtMaintenanceWindowRequestBody.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */

/**
 * Separation of download and install by defining a maintenance window for the installation
 */
export interface MgmtMaintenanceWindowRequestBody {
  /** Schedule for the maintenance window start in quartz cron notation, such as '0 15 10 * * ? 2018'
for 10:15am every day during the year 2018 */
  schedule?: string;
  /** Duration of the window, such as '02:00:00' for 2 hours */
  duration?: string;
  /** A time-zone offset from Greenwich/UTC, such as '+02:00' */
  timezone?: string;
}